Replaced Integer types in favor of Long types for cloudId and branchId parameters

+++ src/main/java/pl/grizzlysoftware/dotykacka/client/v1/api/dto/Employee.java
@@ -25,7 +25,7 @@ public class Employee {;
     @JsonProperty("canonicalname")
     private String canonicalname;
     @JsonProperty("cloudid")
-    private Integer cloudId;
+    private Long cloudId;
     @JsonProperty("deleted")
     private Boolean isDeleted;
     @JsonProperty("email")
