+++ src/main/java/pl/grizzlysoftware/dotykacka/client/v1/api/dto/Tag.java
@@ -20,7 +20,7 @@ public class Tag {
     @JsonProperty("canonicalname")
     public String canonicalName;
     @JsonProperty("cloudid")
-    public Integer cloudId;
+    public Long cloudId;
     @JsonProperty("customerforeignId")
     public Long customerId;
     @JsonProperty("deleted")
