+++ src/main/java/pl/grizzlysoftware/dotykacka/client/v1/api/service/ProductService.java
@@ -33,44 +33,44 @@ import java.util.Collection;
 public interface ProductService {
 
     @GET("{cloudId}/{warehouseId}/list")
-    Call<Collection<ProductWithStockStatus>> getProductsWithStockStatus(@Path("cloudId") Integer cloudId, @Path("warehouseId") long warehouseId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
+    Call<Collection<ProductWithStockStatus>> getProductsWithStockStatus(@Path("cloudId") Long cloudId, @Path("warehouseId") long warehouseId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
 
     @GET("{cloudId}/{productId}")
-    Call<Product> getProduct(@Path("cloudId") Integer cloudId, @Path("productId") Long productId);
+    Call<Product> getProduct(@Path("cloudId") Long cloudId, @Path("productId") Long productId);
 
     @GET("withstockstatus/{cloudId}/{warehouseId}/{productId}")
-    Call<ProductWithStockStatus> getProductWithStockStatus(@Path("cloudId") Integer cloudId, @Path("warehouseId") Long warehouseId, @Path("productId") Long productId);
+    Call<ProductWithStockStatus> getProductWithStockStatus(@Path("cloudId") Long cloudId, @Path("warehouseId") Long warehouseId, @Path("productId") Long productId);
 
     @GET("stockstatus/{cloudId}/{warehouseId}/{productId}")
-    Call<ProductWithStockStatus> getProductStockStatus(@Path("cloudId") Integer cloudId, @Path("warehouseId") Long warehouseId, @Path("productId") Long productId);
+    Call<ProductWithStockStatus> getProductStockStatus(@Path("cloudId") Long cloudId, @Path("warehouseId") Long warehouseId, @Path("productId") Long productId);
 
     @Headers({
             "Content-Type: application/json",
             "Accept: application/json"})
     @POST("{cloudId}/create")
-    Call<Product> createProduct(@Path("cloudId") Integer cloudId, @Body Product product);
+    Call<Product> createProduct(@Path("cloudId") Long cloudId, @Body Product product);
 
     @GET("{cloudId}/{id}/delete")
-    Call<Product> deleteProduct(@Path("cloudId") Integer cloudId, @Path("id") Long id);
+    Call<Product> deleteProduct(@Path("cloudId") Long cloudId, @Path("id") Long id);
 
     @POST("{cloudId}/{id}/update")
     @Headers({
             "Content-Type: application/json",
             "Accept: application/json"})
-    Call<Product> updateProduct(@Path("cloudId") Integer cloudId, @Path("id") Long id, @Body Product consumer);
+    Call<Product> updateProduct(@Path("cloudId") Long cloudId, @Path("id") Long id, @Body Product consumer);
 
     @GET("{cloudId}")
-    Call<Collection<Product>> getProducts(@Path("cloudId") Integer cloudId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
+    Call<Collection<Product>> getProducts(@Path("cloudId") Long cloudId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
 
     @GET("{cloudId}/{productId}/ingredients")
-    Call<Collection<ProductWithStockStatus>> getProductsIngredients(@Path("cloudId") Integer cloudId, @Path("productId") Long productId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
+    Call<Collection<ProductWithStockStatus>> getProductsIngredients(@Path("cloudId") Long cloudId, @Path("productId") Long productId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
 
     @POST("{cloudId}/{productId}/ingredients/edit")
-    Call<Ingredient> createProductIngredient(@Path("cloudId") Integer cloudId, @Body ProductIngredient productIngredient);
+    Call<Ingredient> createProductIngredient(@Path("cloudId") Long cloudId, @Body ProductIngredient productIngredient);
 
     @POST("{cloudId}/{productId}/ingredients/{ingredientId}/delete")
-    Call<Ingredient> deleteProductIngredient(@Path("cloudId") Integer cloudId, @Path("ingredientId") Long ingredientId, @Body ProductIngredient productIngredient);
+    Call<Ingredient> deleteProductIngredient(@Path("cloudId") Long cloudId, @Path("ingredientId") Long ingredientId, @Body ProductIngredient productIngredient);
 
     @GET("{cloudId}/ingredients")
-    Call<Collection<Ingredient>> getIngredients(@Path("cloudId") Integer cloudId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
+    Call<Collection<Ingredient>> getIngredients(@Path("cloudId") Long cloudId, @Query("limit") int limit, @Query("offset") int offset, @Query("sort") String sortBy);
 }
